+++ src/example/java/io/github/computerdaddyguy/jfiletreeprettyprinter/example/LineExtension.java
@@ -1,7 +1,7 @@
 package io.github.computerdaddyguy.jfiletreeprettyprinter.example;
 
 import io.github.computerdaddyguy.jfiletreeprettyprinter.FileTreePrettyPrinter;
-import io.github.computerdaddyguy.jfiletreeprettyprinter.PathUtils;
+import io.github.computerdaddyguy.jfiletreeprettyprinter.PathPredicates;
 import java.nio.file.Path;
 import java.util.function.Function;
 
@@ -9,16 +9,16 @@ public class LineExtension {
 
 	public static void main(String[] args) {
 		Function<Path, String> lineExtension = path -> {
-			if (PathUtils.isDirectory(path) && PathUtils.hasName(path, "api")) {
+			if (PathPredicates.isDirectory(path) && PathPredicates.hasName(path, "api")) {
 				return "\t\t\t// All API code: controllers, etc.";
 			}
-			if (PathUtils.isDirectory(path) && PathUtils.hasName(path, "domain")) {
+			if (PathPredicates.isDirectory(path) && PathPredicates.hasName(path, "domain")) {
 				return "\t\t\t// All domain code: value objects, etc.";
 			}
-			if (PathUtils.isDirectory(path) && PathUtils.hasName(path, "infra")) {
+			if (PathPredicates.isDirectory(path) && PathPredicates.hasName(path, "infra")) {
 				return "\t\t\t// All infra code: database, email service, etc.";
 			}
-			if (PathUtils.isFile(path) && PathUtils.hasName(path, "application.properties")) {
+			if (PathPredicates.isFile(path) && PathPredicates.hasName(path, "application.properties")) {
 				return "\t// Config file";
 			}
 			return null;